requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
Django==4.1.2
django-jazzmin==2.6.0
django-modeltranslation==0.18.7
django-rest-framework==0.1.0
django-rosetta==0.9.8
djangorestframework==3.14.0
idna==3.4
Pillow==9.3.0
polib==1.1.1
pytz==2022.6
requests==2.28.1
sqlparse==0.4.3
typing_extensions==4.4.0
tzdata==2022.6
urllib3==1.26.12
